displays trades on trades page
