+++ Frontend/src/App.tsx
@@ -3,18 +3,40 @@ import HomePage from './pages/Home/HomePage';
 import AnalyticsPage from './pages/Analytics/AnalyticsPage';
 import IssuesPage from './pages/Issues/IssuesPage';
 import TradePage from './pages/TradePage/TradePage';
-import { useState } from 'react';
+import { useState,useEffect } from 'react';
 import { Trade } from './interfaces';
 
 function App() {
   
   const [trades, setTrades] = useState<Trade[]>([]);
   
+  const loadTrades = async () => {
+    const response = await fetch('http://localhost:8080/trades');
+    const jsonData = await response.json();
+    setTrades(jsonData);
+}
+
+  useEffect(() => {
+      loadTrades()
+  },[])
+
+  const addTrade = async (trade: Omit<Trade, 'id'>) => {
+    console.log(trade)
+    const response = await fetch('http://localhost:8080/trades',{
+        method: 'POST',
+        headers: {
+            'Content-Type':'application/json',
+            },
+        body: JSON.stringify(trade),
+        });
+    const newTrade = await response.json();
+    setTrades([...trades, newTrade]);
+  }
  
   const router = createBrowserRouter([
     {
       path: '/',
-      element: <HomePage trades={trades} setTrades={setTrades}/>,
+      element: <HomePage addTrade={addTrade} />,
     },
     {
       path: '/analytics',
