+++ Frontend/src/pages/TradePage/TradeList.tsx
@@ -8,16 +8,20 @@ const TradeList:React.FC<TradeListProps> = ({ trades }) => {
 
     const mappedTrades = trades.map((trade) => {
         return (
-            <li key={trade.name}>
+            <li key={trade.id}>
                 {trade.name}
                 {trade.entryPoint}
+                {trade.exitPoint}
+                {trade.stopLoss}
+                {trade.takeProfit}
+                {trade.tradeDate}
             </li>
         )
     })
     return(
-        <>
+        <ul>
             {mappedTrades}
-        </>
+        </ul>
     )
 }
